refactor: 토픽 이름 상수화
